backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
dashscope>=1.14.0
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.6
